+++ test/ExtendedKeyNode.spec.ts
@@ -53,15 +53,15 @@ describe('ExtendedKeyNode -->', () => {
     describe('constructor should', () => {
 
         it('should create BIP32 node object given no network', () => {
-            const impl = new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[0].key));
-            const neuteredMaster = new ExtendedKeyNode(impl);
+            const node = bip32.fromBase58(extendedKeys.neutered[0].key);
+            const neuteredMaster = new ExtendedKeyNode(node);
 
             expect(neuteredMaster.network.curve).to.be.equal(CurveAlgorithm.secp256k1);
         });
 
         it('should create BIP32 node object given Network.BITCOIN', () => {
-            const impl = new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[0].key));
-            const neuteredMaster = new ExtendedKeyNode(impl, Network.BITCOIN);
+            const node = bip32.fromBase58(extendedKeys.neutered[0].key);
+            const neuteredMaster = new ExtendedKeyNode(node, Network.BITCOIN);
             const nodeBIP32 = neuteredMaster.node as BIP32;
 
             //XXX `BIP32` class cannot be used as a right-hand operator of instanceof
@@ -69,46 +69,46 @@ describe('ExtendedKeyNode -->', () => {
             expect(neuteredMaster.network.curve).to.be.equal(CurveAlgorithm.secp256k1);
 
             // presence checks instead of type check
-            expect(impl.node.privateKey).to.not.be.undefined;
-            expect(impl.node.publicKey).to.not.be.undefined;
-            expect(impl.node.chainCode).to.not.be.undefined;
+            expect(node.privateKey).to.not.be.undefined;
+            expect(node.publicKey).to.not.be.undefined;
+            expect(node.chainCode).to.not.be.undefined;
         });
 
         it('should create NodeEd25519 node object given Network.CATAPULT', () => {
-            const impl = new NodeImpl<NodeEd25519>(NodeEd25519.fromBase58(extendedKeys.neutered[0].key));
-            const neuteredMaster = new ExtendedKeyNode(impl, Network.CATAPULT);
+            const node = NodeEd25519.fromBase58(extendedKeys.neutered[0].key);
+            const neuteredMaster = new ExtendedKeyNode(node, Network.CATAPULT);
             const nodeEd25519 = neuteredMaster.node as NodeEd25519;
 
             expect(nodeEd25519).to.be.instanceof(NodeEd25519);
             expect(neuteredMaster.network.curve).to.be.equal(CurveAlgorithm.ed25519);
-            expect(nodeEd25519.getNetwork().curve).to.be.equal(CurveAlgorithm.ed25519);
+            expect(nodeEd25519.network.curve).to.be.equal(CurveAlgorithm.ed25519);
         });
 
         it('should throw error with inconsistent BIP32 node given network Network.CATAPULT', () => {
             expect((function () {
-                const impl = new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key));
-                const neuteredMaster = new ExtendedKeyNode(impl, Network.CATAPULT);
+                const node = bip32.fromBase58(extendedKeys.neutered[1].key);
+                const neuteredMaster = new ExtendedKeyNode(node, Network.CATAPULT);
             })).to.throw('Given node is of type BIP32 but network does not match Network.BITCOIN.');
         });
 
         it('should throw error with inconsistent NodeEd25519 node given network Network.BITCOIN', () => {
             expect((function () {
-                const impl = new NodeImpl<NodeEd25519>(NodeEd25519.fromBase58(extendedKeys.neutered[1].key));
-                const neuteredMaster = new ExtendedKeyNode(impl, Network.BITCOIN);
+                const node = NodeEd25519.fromBase58(extendedKeys.neutered[1].key);
+                const neuteredMaster = new ExtendedKeyNode(node, Network.BITCOIN);
             })).to.throw('Given node is of type NodeEd25519 but network does not match Network.CATAPULT.');
         });
 
         it('create master key with payload for "m" path', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[0].key)));
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[0].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[0].key));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[0].key));
 
             expect(neuteredMaster.isMaster()).to.be.equal(true);
             expect(nonNeuteredMaster.isMaster()).to.be.equal(true);
         });
 
         it('create child key with payload for "m/0" path', () => {
-            const neuteredChild = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
-            const nonNeuteredChild = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[1].key)));
+            const neuteredChild = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
+            const nonNeuteredChild = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[1].key));
 
             expect(neuteredChild.isMaster()).to.be.equal(false);
             expect(nonNeuteredChild.isMaster()).to.be.equal(false);
@@ -117,7 +117,7 @@ describe('ExtendedKeyNode -->', () => {
         it('create neutered from neutered keys', () => {
             extendedKeys.neutered.map(
                 (neuteredKey) => {
-                    const neuteredNode = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(neuteredKey.key)));
+                    const neuteredNode = new ExtendedKeyNode(bip32.fromBase58(neuteredKey.key));
                     expect(neuteredNode.isNeutered()).to.be.equal(true);
                 });
         });
@@ -125,7 +125,7 @@ describe('ExtendedKeyNode -->', () => {
         it('create non-neutered from non-neutered keys', () => {
             extendedKeys.nonNeutered.map(
                 (nonNeuteredKey) => {
-                    const nonNeuteredNode = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(nonNeuteredKey.key)));
+                    const nonNeuteredNode = new ExtendedKeyNode(bip32.fromBase58(nonNeuteredKey.key));
                     expect(nonNeuteredNode.isNeutered()).to.be.equal(false);
                 });
         });
@@ -199,22 +199,22 @@ describe('ExtendedKeyNode -->', () => {
 
     describe('getPublicNode() should', () => {
         it('create neutered from non-neutered', () => {
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[0].key)));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[0].key));
             const publicNode = nonNeuteredMaster.getPublicNode();
 
             expect(publicNode.isNeutered()).to.be.equal(true);
         });
 
         it('create neutered from neutered', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
             const publicNode = neuteredMaster.getPublicNode();
 
             expect(publicNode.isNeutered()).to.be.equal(true);
         });
 
         it('use correct network after being neutered', () => {
-            const impl = new NodeImpl<NodeEd25519>(NodeEd25519.fromBase58(extendedKeys.neutered[1].key));
-            const neuteredMaster = new ExtendedKeyNode(impl, Network.CATAPULT);
+            const node = NodeEd25519.fromBase58(extendedKeys.neutered[1].key);
+            const neuteredMaster = new ExtendedKeyNode(node, Network.CATAPULT);
             const publicNode = neuteredMaster.getPublicNode();
 
             // check that Network.CATAPULT was used correctly
@@ -229,8 +229,8 @@ describe('ExtendedKeyNode -->', () => {
 
     describe('toBase58() should', () => {
         it('produce same payloads', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[1].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[1].key));
             const neuteredBase58 = neuteredMaster.toBase58();
             const nonNeuteredBase58 = nonNeuteredMaster.toBase58();
 
@@ -242,13 +242,13 @@ describe('ExtendedKeyNode -->', () => {
     describe('getPrivateKey() should', () => {
         it('should throw error with neutered nodes', () => {
             expect((function () {
-                const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
+                const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
                 const privateKey = neuteredMaster.getPrivateKey();
             })).to.throw('Cannot read private key out of extended public key.');
         });
 
         it('should return hexadecimal notation by default', () => {
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[1].key)));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[1].key));
             const privateKey = nonNeuteredMaster.getPrivateKey();
 
             expect(privateKey.length).to.be.equal(64);
@@ -256,7 +256,7 @@ describe('ExtendedKeyNode -->', () => {
         });
 
         it('should return binary notation with ENC_BIN', () => {
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[1].key)));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[1].key));
             const privateKey: Buffer = nonNeuteredMaster.getPrivateKey(KeyEncoding.ENC_BIN) as Buffer;
             const uintArray = new Uint8Array(privateKey);
 
@@ -280,8 +280,8 @@ describe('ExtendedKeyNode -->', () => {
 
     describe('getPublicKey() should', () => {
         it('should return public key for both neutered and non-neutered', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[0].key)));
-            const nonNeuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.nonNeutered[0].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[0].key));
+            const nonNeuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.nonNeutered[0].key));
             const neuteredPubKey = neuteredMaster.getPublicKey();
             const nonNeuteredPubKey = nonNeuteredMaster.getPublicKey();
 
@@ -290,7 +290,7 @@ describe('ExtendedKeyNode -->', () => {
         });
 
         it('should return hexadecimal notation by default', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
             const publicKey = neuteredMaster.getPublicKey();
 
             expect(publicKey.length).to.be.equal(64);
@@ -298,7 +298,7 @@ describe('ExtendedKeyNode -->', () => {
         });
 
         it('should return binary notation with ENC_BIN', () => {
-            const neuteredMaster = new ExtendedKeyNode(new NodeImpl<BIP32>(bip32.fromBase58(extendedKeys.neutered[1].key)));
+            const neuteredMaster = new ExtendedKeyNode(bip32.fromBase58(extendedKeys.neutered[1].key));
             const publicKey: Buffer = neuteredMaster.getPublicKey(KeyEncoding.ENC_BIN) as Buffer;
             const uintArray = new Uint8Array(publicKey);